backend/pyproject.toml
[tool.poetry]
name = "voice-assistant"
version = "0.1.0"
description = ""
authors = ["ruizguille <guillermo@codeawake.com>"]
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.12"
python-dotenv = "^1.0.1"
groq = "^0.7.0"
deepgram-sdk = "^3.2.7"
rich = "^13.7.1"
requests = "^2.32.2"
pyaudio = "^0.2.14"
fastapi = "^0.111.0"
uvicorn = {extras = ["standard"], version = "^0.29.0"}
pydantic = "^2.7.1"
pydantic-settings = "^2.2.1"

httpx = "^0.27.0"
[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
local = "app.local_assistant:main"